requirements.txt
pygame-ce>=2.4.0
pytest>=8.0.0
